paging: integrate paging with the log crate (#13)
paging: integrate paging with the log crate

- Utilize the globally configured logger for paging operations(for
no_std).
- Add a simple logger implementation for testing purposes (for std).
```
---------------------------------------------[0x0000000000000000 0x0000000000007FFF]------------------------------------------------
                                                      6362        52 51                                   12 11 9 8 7 6 5 4 3 2 1 0
                                                      |N|           |                                        |   |M|M|I| |P|P|U|R| |
                                                      |X| Available |     Page-Map Level-4 Base Address      |AVL|B|B|G|A|C|W|/|/|P|
                                                      | |           |                                        |   |Z|Z|N| |D|T|S|W| |
------------------------------------------------------------------------------------------------------------------------------------
PML4 |  [0x0000000000000000 0x0000000000007FFF]       |0|00000000000|0000000000011001001110001110011101001101|000|0|0|0|0|0|0|1|1|1|
PDP  |    [0x0000000000000000 0x0000000000007FFF]     |0|00000000000|0000000000011001001110001110011101001110|000|0|0|0|0|0|0|1|1|1|
PD   |      [0x0000000000000000 0x0000000000007FFF]   |0|00000000000|0000000000011001001110001110011101001111|000|0|0|0|0|0|0|1|1|1|
PT   |        [0x0000000000000000 0x0000000000000FFF] |0|00000000000|0000000000000000000000000000000000000000|000|0|0|0|0|0|0|1|0|1|
PT   |        [0x0000000000001000 0x0000000000001FFF] |0|00000000000|0000000000000000000000000000000000000001|000|0|0|0|0|0|0|1|0|1|
PT   |        [0x0000000000002000 0x0000000000002FFF] |0|00000000000|0000000000000000000000000000000000000010|000|0|0|0|0|0|0|1|0|1|
PT   |        [0x0000000000003000 0x0000000000003FFF] |0|00000000000|0000000000000000000000000000000000000011|000|0|0|0|0|0|0|1|0|1|
PT   |        [0x0000000000004000 0x0000000000004FFF] |0|00000000000|0000000000000000000000000000000000000100|000|0|0|0|0|0|0|1|0|1|
PT   |        [0x0000000000005000 0x0000000000005FFF] |0|00000000000|0000000000000000000000000000000000000101|000|0|0|0|0|0|0|1|0|1|
PT   |        [0x0000000000006000 0x0000000000006FFF] |0|00000000000|0000000000000000000000000000000000000110|000|0|0|0|0|0|0|1|0|1|
PT   |        [0x0000000000007000 0x0000000000007FFF] |0|00000000000|0000000000000000000000000000000000000111|000|0|0|0|0|0|0|1|0|1|
------------------------------------------------------------------------------------------------------------------------------------
```

+++ src/aarch64/structs.rs
@@ -1,10 +1,10 @@
+use crate::{MemoryAttributes, PtError, PtResult};
+use alloc::string::String;
 use bitfield_struct::bitfield;
 use core::{
     fmt::{self, Display, Formatter},
     ops::{Add, Sub},
 };
-
-use crate::{MemoryAttributes, PtError, PtResult};
 
 pub(crate) const MAX_VA: u64 = 0x0000_ffff_ffff_ffff;
 
@@ -73,7 +73,6 @@ impl VMSAv864TableDescriptor {
                 panic!("allocated page is not 4k aligned {:X}", next_level_table_base);
             }
 
-            // println!("next_level_table_base: {:X}", next_level_table_base);
             let nlta_lower = (next_level_table_base & PAGE_MAP_ENTRY_PAGE_TABLE_BASE_ADDRESS_LOWER_MASK)
                 >> PAGE_MAP_ENTRY_PAGE_TABLE_BASE_ADDRESS_LOWER_SHIFT;
             let nlta_upper = ((next_level_table_base & PAGE_MAP_ENTRY_PAGE_TABLE_BASE_ADDRESS_UPPER_MASK)
@@ -126,7 +125,6 @@ impl VMSAv864TableDescriptor {
         self.set_valid_desc(false);
     }
 
-    #[cfg(test)]
     pub fn dump_entry(&self) -> String {
         let valid_desc = self.valid_desc() as u64;
         let table_desc = self.table_desc() as u64;
@@ -134,7 +132,7 @@ impl VMSAv864TableDescriptor {
         let nlta_upper = self.nlta_upper() as u64;
         let access_flag = self.access_flag() as u64;
         let ignored1 = self.ignored1() as u64;
-        let nlta_lower = self.nlta_lower() as u64;
+        let nlta_lower = self.nlta_lower();
         let ignored2 = self.ignored2() as u64;
         let ignored3 = self.ignored3() as u64;
         let pxn_table = self.pxn_table() as u64;
@@ -324,7 +322,6 @@ impl VMSAv864PageDescriptor {
         self.set_descriptor_type(0);
     }
 
-    #[cfg(test)]
     pub fn dump_entry(&self) -> String {
         let descriptor_type = self.descriptor_type() as u64;
         let attribute_index = self.attribute_index() as u64;
@@ -408,7 +405,6 @@ impl Sub<u64> for PageLevel {
     }
 }
 
-#[cfg(test)]
 impl fmt::Display for PageLevel {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         let level_name = match self {
